Adopt compose logic over inheritance
