+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/transactionhandler/AbstractEntityCrudTransactionHandler.java
@@ -24,6 +24,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import com.hedera.mirror.common.domain.contract.ContractResult;
 import com.hedera.mirror.common.domain.entity.AbstractEntity;
 import com.hedera.mirror.common.domain.entity.EntityId;
 import com.hedera.mirror.common.domain.entity.EntityOperation;
@@ -39,6 +40,11 @@ abstract class AbstractEntityCrudTransactionHandler<T extends AbstractEntity> im
 
     @Getter
     private final TransactionType type;
+
+    @Override
+    public ContractResult getContractResult(Transaction transaction, RecordItem recordItem) {
+        return null;
+    }
 
     @Override
     public final void updateTransaction(Transaction transaction, RecordItem recordItem) {
