+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/transactionhandler/ContractUpdateTransactionHandler.java
@@ -58,7 +58,6 @@ class ContractUpdateTransactionHandler extends AbstractEntityCrudTransactionHand
     }
 
     // We explicitly ignore the updated fileID field since hedera nodes do not allow changing the bytecode after create
-    @SuppressWarnings("java:S1874")
     @Override
     protected void doUpdateEntity(Contract contract, RecordItem recordItem) {
         var transactionBody = recordItem.getTransactionBody().getContractUpdateInstance();
