+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/transactionhandler/CryptoAdjustAllowanceTransactionHandler.java
@@ -26,7 +26,9 @@ import com.hederahashgraph.api.proto.java.TokenAllowance;
 import java.util.List;
 import javax.inject.Named;
 
+import com.hedera.mirror.common.domain.contract.ContractResult;
 import com.hedera.mirror.common.domain.transaction.RecordItem;
+import com.hedera.mirror.common.domain.transaction.Transaction;
 import com.hedera.mirror.common.domain.transaction.TransactionType;
 import com.hedera.mirror.importer.parser.record.entity.EntityListener;
 
@@ -35,6 +37,11 @@ class CryptoAdjustAllowanceTransactionHandler extends AbstractAllowanceTransacti
 
     public CryptoAdjustAllowanceTransactionHandler(EntityListener entityListener) {
         super(entityListener);
+    }
+
+    @Override
+    public ContractResult getContractResult(Transaction transaction, RecordItem recordItem) {
+        return null;
     }
 
     @Override
