+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/transactionhandler/CryptoApproveAllowanceTransactionHandler.java
@@ -26,6 +26,7 @@ import com.hederahashgraph.api.proto.java.TokenAllowance;
 import java.util.List;
 import javax.inject.Named;
 
+import com.hedera.mirror.common.domain.contract.ContractResult;
 import com.hedera.mirror.common.domain.transaction.RecordItem;
 import com.hedera.mirror.common.domain.transaction.Transaction;
 import com.hedera.mirror.common.domain.transaction.TransactionType;
@@ -36,6 +37,11 @@ class CryptoApproveAllowanceTransactionHandler extends AbstractAllowanceTransact
 
     public CryptoApproveAllowanceTransactionHandler(EntityListener entityListener) {
         super(entityListener);
+    }
+
+    @Override
+    public ContractResult getContractResult(Transaction transaction, RecordItem recordItem) {
+        return null;
     }
 
     @Override
