+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/transactionhandler/CryptoUpdateTransactionHandler.java
@@ -42,7 +42,6 @@ class CryptoUpdateTransactionHandler extends AbstractEntityCrudTransactionHandle
     }
 
     @Override
-    @SuppressWarnings("java:S1874")
     protected void doUpdateEntity(Entity entity, RecordItem recordItem) {
         var transactionBody = recordItem.getTransactionBody().getCryptoUpdateAccount();
