+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/transactionhandler/UnknownDataTransactionHandler.java
@@ -22,12 +22,19 @@ package com.hedera.mirror.importer.parser.record.transactionhandler;
 
 import javax.inject.Named;
 
+import com.hedera.mirror.common.domain.contract.ContractResult;
 import com.hedera.mirror.common.domain.entity.EntityId;
-import com.hedera.mirror.common.domain.transaction.TransactionType;
 import com.hedera.mirror.common.domain.transaction.RecordItem;
+import com.hedera.mirror.common.domain.transaction.Transaction;
+import com.hedera.mirror.common.domain.transaction.TransactionType;
 
 @Named
 class UnknownDataTransactionHandler implements TransactionHandler {
+
+    @Override
+    public ContractResult getContractResult(Transaction transaction, RecordItem recordItem) {
+        return null;
+    }
 
     @Override
     public EntityId getEntity(RecordItem recordItem) {
