+++ hedera-mirror-importer/src/test/java/com/hedera/mirror/importer/parser/record/entity/TransactionSignatureTest.java
@@ -63,6 +63,7 @@ import com.hedera.mirror.common.domain.transaction.RecordItem;
 import com.hedera.mirror.common.domain.transaction.TransactionSignature;
 import com.hedera.mirror.common.domain.transaction.TransactionType;
 import com.hedera.mirror.importer.addressbook.AddressBookService;
+import com.hedera.mirror.importer.domain.EntityIdService;
 import com.hedera.mirror.importer.parser.CommonParserProperties;
 import com.hedera.mirror.importer.parser.record.NonFeeTransferExtractionStrategy;
 import com.hedera.mirror.importer.parser.record.RecordParserProperties;
@@ -83,6 +84,9 @@ class TransactionSignatureTest {
 
     @Mock
     private EntityListener entityListener;
+
+    @Mock
+    private EntityIdService entityIdService;
 
     @Mock
     private NonFeeTransferExtractionStrategy nonFeeTransferExtractionStrategy;
