+++ hedera-mirror-importer/src/test/java/com/hedera/mirror/importer/parser/record/transactionhandler/TokenBurnTransactionHandlerTest.java
@@ -32,7 +32,7 @@ class TokenBurnTransactionHandlerTest extends AbstractTransactionHandlerTest {
 
     @Override
     protected TransactionHandler getTransactionHandler() {
-        return new TokenBurnTransactionHandler(entityIdService, entityListener, entityProperties);
+        return new TokenBurnTransactionHandler(entityIdService, entityProperties);
     }
 
     @Override
